add getConfusablesExplanation
