+++ app/util/confusables/index.js
@@ -1,4 +1,6 @@
 import { confusables } from 'unicode-confusables';
+import { strings } from '../../../locales/i18n';
+import confusablesMap from 'unicode-confusables/data/confusables.json';
 
 export const collectConfusables = (ensName) => {
 	const key = 'similarTo';
@@ -19,3 +21,14 @@ const zeroWidthPoints = new Set([
 ]);
 
 export const hasZeroWidthPoints = (char) => zeroWidthPoints.has(char);
+
+export const getConfusablesExplanation = (confusableCollection) =>
+	confusableCollection
+		.map((confusable) => {
+			const key = confusable;
+			const value = confusablesMap[confusable];
+			return hasZeroWidthPoints(key)
+				? strings('transaction.contains_zero_width')
+				: `'${key}' ${strings('transaction.similar_to')} '${value}'`;
+		})
+		.reduce((unique, item) => (unique.includes(item) ? unique : [...unique, item]), []);
